requirements.txt
agentic-doc
openai
groq
fastapi
uvicorn
python-dotenv
python-multipart
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib 
gspread
beautifulsoup4
langgraph